Improve notification format for .deb files.
If we have build log and debdiff files matching a given .deb, include
them all on a single line instead of listing them across several lines.
This should make it easier to pick the .deb file to download and run
tests against, while still providing links to the associated files.