Implement monitoring repositories / packages.
This gives us peace of mind since we'll be notified about the updates
in Debian/Raspberry OS for the packages we need to be careful about.

On the Debian side, an updated raspi-firmware would be problematic but
that seems unlikely to happen. Let's monitor bookworm-security via the
(almost-)inevitable copy that ends up in bookworm-proposed-updates.

On the Raspberry OS side, an updated firmware-brcm80211 seems possible,
and updated linux-image-* packages seem very likely. We will want to
check then merge all of those!