Sync the suite directory explicitly.
That duplicates the ppa.suite parameter, but that makes it much easier
and less riskier than sync-ing the whole directory, when it's time to
add --delete (see previous commit).

This can be reconsidered when it's time to focus on trixie instead of
(or in addition to) bookworm, which might or might not mean having two
different suite directories to deal with.